refactor(gotest): reorganize assertions, utils, and t.go
Move assertions into alphabetical order in assertions.go, extract
Must and Each into utils.go, and trim t.go to struct and scoping methods.
